Fixed up the damage/hitting ratio.  Combat feels slightly more realistic, and it almost feels like a real game you could play! Next up, making the messaging system more robust.

+++ src/character_classes/character.h
@@ -704,7 +704,7 @@ class Character
          * The character's accuracy, representing the percentage
          * of how often a character hits something.
          */
-        float accuracy_stat();
+        int accuracy_stat();
 
 
         /**
